[MemProf] Split MemProfiler into Instrumentation and Use.
