+++ graphite_command_reference.md
@@ -1,0 +1,1 @@
+../../support/graphite_command_reference.md
